LayoutTests/js/dom/array-sort-non-writeable-element-expected.txt

PASS array.sort((a, b) => a - b) threw exception TypeError: Attempted to assign to readonly property..
PASS successfullyParsed is true

TEST COMPLETE
